requirements.txt
flask
gunicorn
openai
pyairtable
python-dotenv==3.0.2
stripe==7.9.0
